feat: add scentDesc 중복제거
