+++ app.js
@@ -1,9 +1,12 @@
 const http = require('http')
 const express = require('express')
 const routes = require("./routes");
-
+const cors = require('cors');
 const app = express()
 app.use(express.json())
+app.use(cors({
+  origin: '*'
+}));
 app.use(routes)
 
 app.get('/',(req,res) => {
