+++ models/productDao.js
@@ -1,9 +1,10 @@
 const { appDataSource } = require("./datasource")
 const {updateQuery} = require("../util/update")
+const {scentDesc} = require("../util/scentDesc")
 const getProducts = async (brandName,scentName) =>{
     const result =  await appDataSource.query(`
     SELECT products.id,products.name as product_name,products.price,products.description,products.brand_id,products.scent_id,products.stock,
-    scents.name,scents.scent_desc,
+    scents.name as scent_name,scents.scent_desc,
     brands.name as brand_name,
     images.image_source 
     FROM products 
@@ -11,23 +12,8 @@ const getProducts = async (brandName,scentName) =>{
     LEFT JOIN scents ON products.scent_id = scents.id
     LEFT JOIN brands ON products.brand_id = brands.id 
     WHERE products.brand_id = '${brandName}' or products.scent_id = '${scentName}'`)
-    
-    const scentDescMap = {};
-    
-    result.forEach((product) => {
-        const { scent_desc, ...rest } = product;
-        if (!(scent_desc in scentDescMap)) {
-          scentDescMap[scent_desc] = { scent_desc, products: [] };
-        }
-        scentDescMap[scent_desc].products.push(rest);
-      });
-    
-      const uniqueProducts = Object.values(scentDescMap).map(({ scent_desc, products }) => ({
-        scent_desc,
-        products,
-      }));
-    
-      return uniqueProducts;
+    const scentDescMap = await scentDesc(result)
+      return scentDescMap
 }
